system-integration/class-09-05/transactions_api/requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
flask-redis==0.4.0
python-dotenv==0.19.0